+++ Modules/votingFrame.lua
@@ -1770,7 +1770,8 @@ end
 --@param noAutopass: true or false or nil. Determine whether we force no autopass.
 --@param announceInChat: true or false or nil. Determine if the reannounce sessions should be announced in chat.
 function RCVotingFrame:ReannounceOrRequestRoll(namePred, sesPred, isRoll, noAutopass, announceInChat)
-	addon.Log:D("ReannounceOrRequestRoll", namePred, sesPred, isRoll, noAutopass, announceInChat)
+	addon.Log:D("ReannounceOrRequestRoll", namePred, sesPred, isRoll, noAutopass, announceInChat,
+	MSA_DROPDOWNMENU_MENU_VALUE)
 	local rerollTable = TempTable:Acquire()
 	local changeResponseData = TempTable:Acquire()
 	local rollsData = TempTable:Acquire()